requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=3.0.0
pymongo>=4.0.0
email-validator>=1.1.3
bcrypt==4.0.1
passlib==1.7.4
pydantic[email]>=1.8.2
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5 